multi_agent_doc_system/requirements.txt
pydantic-ai
lightrag
openai
python-dotenv
PyPDF2
pdfminer.six
unstructured
reportlab
pillow-heif
